* avr-dis.c (avr_operand): Warning fix.

+++ opcodes/avr-dis.c
@@ -142,7 +142,7 @@ avr_operand (unsigned int insn, unsigned int insn2, unsigned int pc, int constra
       /* See PR binutils/2545.  Ideally we would like to display the hex
 	 value of the address only once, but this would mean recoding
 	 objdump_print_address() which would affect many targets.  */
-      sprintf (buf, "%#lx", * sym_addr);      
+      sprintf (buf, "%#lx", (unsigned long ) *sym_addr);      
       sprintf (comment, "0x");
 
       break;
